requirements.txt
google-cloud-texttospeech
google-auth
streamlit